+++ CHANGELOG.md
@@ -1,10 +1,17 @@
-<a name="2.6.3"></a>
+<a name="2.6.4"></a>
 
-# [2.6.3](https://gitlab.com/atjdev/ezio-api/-/merge_requests/143) (2022-04-28)
+# [2.6.4](https://gitlab.com/atjdev/ezio-api/-/merge_requests/143) (2022-04-29)
 
 ### Bug Fixes
 
 - [**M015**] **Transport-Task**: spliting update TT auto from create.
+
+# [2.6.3](https://gitlab.com/atjdev/ezio-api/-/merge_requests/142) (2022-04-28)
+
+### Features
+
+- [**M014**] **warping**: add some queries for AT4 form.
+- add ifs shop order operation entity and resolver.
 
 <a name="2.6.2"></a>